+++ src/main/java/io/github/astrapi69/swing/action/ToggleFullScreenAction.java
@@ -27,11 +27,13 @@ package io.github.astrapi69.swing.action;
 import java.awt.Frame;
 import java.awt.GraphicsDevice;
 import java.awt.event.ActionEvent;
+import java.io.Serial;
 
 import javax.swing.AbstractAction;
 import javax.swing.JFrame;
 
 import lombok.AccessLevel;
+import lombok.NonNull;
 import lombok.experimental.FieldDefaults;
 
 /**
@@ -42,6 +44,7 @@ public class ToggleFullScreenAction extends AbstractAction
 {
 
 	/** The Constant serialVersionUID. */
+	@Serial
 	private static final long serialVersionUID = 1L;
 
 	/** The frame. */
@@ -67,16 +70,24 @@ public class ToggleFullScreenAction extends AbstractAction
 	@Override
 	public void actionPerformed(final ActionEvent actionEvent)
 	{
-		onToggleFullScreen(actionEvent);
+		onToggleFullScreen();
 	}
 
 	/**
 	 * Callback method to interact on toggle to full screen
-	 *
-	 * @param actionEvent
-	 *            the action event
 	 */
-	protected void onToggleFullScreen(final ActionEvent actionEvent)
+	protected void onToggleFullScreen()
+	{
+		ToggleFullScreenAction.toggleFullScreen(frame);
+	}
+
+	/**
+	 * Toggle to full screen and back
+	 * 
+	 * @param frame
+	 *            the {@link JFrame} object
+	 */
+	public static void toggleFullScreen(final @NonNull JFrame frame)
 	{
 		GraphicsDevice device = frame.getGraphicsConfiguration().getDevice();
 		frame.setExtendedState(Frame.MAXIMIZED_BOTH);