+++ src/main/java/io/github/astrapi69/swing/menu/enumeration/MenuType.java
@@ -68,5 +68,10 @@ public enum MenuType
 	/**
 	 * Represents the popup menu type
 	 */
-	POPUP
+	POPUP,
+
+	/**
+	 * if the menu type is unknown
+	 */
+	UNKNOWN
 }