+++ src/main/java/io/github/astrapi69/swing/menu/factory/JMenuFactory.java
@@ -29,7 +29,6 @@ import java.awt.event.ActionListener;
 import javax.swing.JMenu;
 
 import io.github.astrapi69.swing.menu.MenuExtensions;
-import io.github.astrapi69.swing.menu.model.MenuInfo;
 import io.github.astrapi69.swing.menu.model.MenuItemInfo;
 import lombok.NonNull;
 
@@ -48,8 +47,7 @@ public class JMenuFactory
 	 */
 	public static JMenu newJMenu(final @NonNull String text)
 	{
-		return MenuItemInfo.builder().menuInfo(MenuInfo.builder().text(text).build()).build()
-			.toJMenu();
+		return MenuItemInfo.builder().text(text).build().toJMenu();
 	}
 
 	/**
@@ -75,8 +73,7 @@ public class JMenuFactory
 	 */
 	public static JMenu newJMenu(final @NonNull String text, final int mnemonic)
 	{
-		return MenuItemInfo.builder()
-			.menuInfo(MenuInfo.builder().text(text).mnemonic(mnemonic).build()).build().toJMenu();
+		return MenuItemInfo.builder().text(text).mnemonic(mnemonic).build().toJMenu();
 	}
 
 	/**
@@ -93,9 +90,8 @@ public class JMenuFactory
 	public static JMenu newJMenu(final @NonNull String text, final int mnemonic,
 		final @NonNull ActionListener actionListener)
 	{
-		return MenuItemInfo.builder()
-			.menuInfo(MenuInfo.builder().text(text).mnemonic(mnemonic).build())
-			.actionListener(actionListener).build().toJMenu();
+		return MenuItemInfo.builder().text(text).mnemonic(mnemonic).actionListener(actionListener)
+			.build().toJMenu();
 	}
 
 	/**
@@ -109,9 +105,7 @@ public class JMenuFactory
 	 */
 	public static JMenu newJMenu(final @NonNull String text, final char mnemonic)
 	{
-		return MenuItemInfo.builder()
-			.menuInfo(
-				MenuInfo.builder().text(text).mnemonic(MenuExtensions.toMnemonic(mnemonic)).build())
+		return MenuItemInfo.builder().text(text).mnemonic(MenuExtensions.toMnemonic(mnemonic))
 			.build().toJMenu();
 	}
 
@@ -129,9 +123,7 @@ public class JMenuFactory
 	public static JMenu newJMenu(final @NonNull String text, final char mnemonic,
 		final @NonNull ActionListener actionListener)
 	{
-		return MenuItemInfo.builder()
-			.menuInfo(
-				MenuInfo.builder().text(text).mnemonic(MenuExtensions.toMnemonic(mnemonic)).build())
+		return MenuItemInfo.builder().text(text).mnemonic(MenuExtensions.toMnemonic(mnemonic))
 			.actionListener(actionListener).build().toJMenu();
 	}