+++ src/main/java/io/github/astrapi69/swing/menu/model/KeyStrokeInfo.java
@@ -26,6 +26,7 @@ package io.github.astrapi69.swing.menu.model;
 
 import javax.swing.KeyStroke;
 
+import io.github.astrapi69.swing.menu.KeyStrokeExtensions;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -51,10 +52,6 @@ import lombok.experimental.SuperBuilder;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class KeyStrokeInfo
 {
-	/**
-	 * the {@link Character} object with the value for a keyboard key
-	 */
-	Character keyChar;
 
 	/**
 	 * the {@link Integer} object that specifies the numeric code for a keyboard key
@@ -73,8 +70,7 @@ public class KeyStrokeInfo
 	Boolean onKeyRelease;
 
 	/**
-	 * the {@link String} object that specifies the keystroke.
-	 *
+	 * the {@link String} object that specifies the keystroke. <br/>
 	 * For more information @see KeyStroke#getKeyStroke(String)
 	 */
 	String keystrokeAsString;
@@ -89,10 +85,7 @@ public class KeyStrokeInfo
 	 */
 	public static KeyStrokeInfo toKeyStrokeInfo(final @NonNull KeyStroke keyStroke)
 	{
-		return KeyStrokeInfo.builder().keyCode(keyStroke.getKeyCode())
-			.keyChar(keyStroke.getKeyChar()).modifiers(keyStroke.getModifiers())
-			.onKeyRelease(keyStroke.isOnKeyRelease()).keystrokeAsString(keyStroke.toString())
-			.build();
+		return KeyStrokeExtensions.toKeyStrokeInfo(keyStroke);
 	}
 
 	/**
@@ -104,7 +97,6 @@ public class KeyStrokeInfo
 	public KeyStrokeInfo set(final @NonNull KeyStroke keyStroke)
 	{
 		this.keyCode = keyStroke.getKeyCode();
-		this.keyChar = keyStroke.getKeyChar();
 		this.modifiers = keyStroke.getModifiers();
 		this.onKeyRelease = keyStroke.isOnKeyRelease();
 		this.keystrokeAsString = keyStroke.toString();
@@ -146,25 +138,6 @@ public class KeyStrokeInfo
 				{
 					return keyStroke;
 				}
-			}
-		}
-
-		if (keyChar != null)
-		{
-
-			if (modifiers != null)
-			{
-				keyStroke = KeyStroke.getKeyStroke(keyChar, modifiers.intValue());
-				if (keyStroke != null)
-				{
-					return keyStroke;
-				}
-			}
-
-			keyStroke = KeyStroke.getKeyStroke(keyChar);
-			if (keyStroke != null)
-			{
-				return keyStroke;
 			}
 		}
 		return keyStroke;