+++ src/main/java/io/github/astrapi69/swing/menu/model/MenuInfo.java
@@ -96,6 +96,12 @@ public class MenuInfo
 	String relativeToMenuId;
 
 	/**
+	 * The action command or the fully qualified class name of the action listener for this menu
+	 * component
+	 */
+	String actionCommand;
+
+	/**
 	 * Factory method that creates a {@link MenuItemInfo} object from this {@link MenuInfo} object
 	 * and the given {@link ActionListener} object
 	 *